pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tablement"
version = "0.1.0"
description = "Resy reservation sniper for accessibility"
requires-python = ">=3.11"
dependencies = [
    "httpx[http2]>=0.27",
    "pydantic>=2.5",
    "pyyaml>=6.0",
    "click>=8.1",
    "keyring>=25.0",
    "rich>=13.0",
    "ntplib>=0.4",
    "orjson>=3.9",
]

[project.optional-dependencies]
web = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "jinja2>=3.1",
    "sse-starlette>=2.0",
    "anthropic>=0.42",
    "supabase>=2.0",
    "PyJWT>=2.8",
    "stripe>=8.0",
    "apscheduler>=3.10,<4.0",
    "cryptography>=42.0",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.1",
    "respx>=0.21",
    "time-machine>=2.13",
    "ruff>=0.2",
]

[project.scripts]
tablement = "tablement.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/tablement"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
